chore: rename to genric names instead of poll
