+++ packages/workers/snapshot-relay/src/index.ts
@@ -1,7 +1,7 @@
 import { createCors, error, Router } from 'itty-router';
 
 import createPoll from './handlers/createPoll';
-import getPoll from './handlers/getPoll';
+import getProposal from './handlers/getProposal';
 import votePoll from './handlers/votePoll';
 import type { Env } from './types';
 
@@ -14,8 +14,8 @@ const router = Router();
 
 router.all('*', preflight);
 router.get('/', () => new Response('Snapshot Relay'));
-router.get('/getPoll/:id/:voter', ({ params }) =>
-  getPoll(params.id, params.voter)
+router.get('/getProposal/:id/:voter', ({ params }) =>
+  getProposal(params.id, params.voter)
 );
 router.post('/createPoll', createPoll);
 router.post('/votePoll', votePoll);
